% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demToCSV.R
\name{demToCSV}
\alias{demToCSV}
\title{Convert a Raster DEM file to a Useable CSV that can be inputted into ClimateNA}
\usage{
demToCSV(file, outdir = dirname(file), srs = "EPSG:4326")
}
\arguments{
\item{file}{string. A TIFF file at the desired spatial resolution.}

\item{outdir}{The output location for the formatted CSV file. The default is the same location as input.}

\item{srs}{string. A valid WKT string or SRS definition, such as "EPSG:4326" or "ESRI:102761" or NULL. If the DEM is already in lat/long projection, use NULL}
}
\value{
A CSV file representing the inputted DEM
}
\description{
Input a TIFF file containing the DEM elevation values clipped to the area of focus. This tool will convert and produce a point featureclass shapefile with an identical name and use that shapefile's points and point coordinates to produce a CSV file. This CSV file is formatted so it can be inputted into ClimateNA.
}
\examples{
#files <- list.files(pattern='*.tif$')
#roiDEM <- files[3]
#demToCSV(roiDEM)
}
\author{
Michael Burnett - UBC
}
