% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histClimateNA.R
\name{histClimateNA}
\alias{histClimateNA}
\title{Access Historical ClimateNA Data}
\usage{
histClimateNA(file, dateR, tFrame, exe, outdir = dirname(file))
}
\arguments{
\item{file}{string. The prepared DEM as a CSV file.}

\item{dateR}{string. The historical year (or time frame) for the climate data. If annual data is desired, make sure it is in the number format of 'XXXX'. If decadal or 30 years normals are desired, make sure it is in the 10 years format of 'XXX1_XXX0'. Underscore is necessary, as opposed to dash.}

\item{tFrame}{string. The averaged time frame of each climate variable. Use 'M' for monthly, 'Y' for annual, and 'S' for seasonal.}

\item{exe}{string. Full name of the ClimateNA exe file. The working directory must be the location of this file.}

\item{outdir}{The output directory for the climate data. As a default, it will be outputted to the same location as the input CSV file.}
}
\value{
A CSV file with historical climate data
}
\description{
This is a specialty function that has several preliminary requirements. The output produces the  historical climate variables for the provided CSV file.
In order to properly use this function, you must prepare the CSV file so that it is compatible with the ClimateNA requirements (using the demToCSV tool).
The working directory must be set to the location which the ClimateNA app is located.
}
\examples{
## Annual climate data
#setwd("E:/Climatena_v721");getwd() # Set up location for the application
#exe <- "E:/Climatena_v721/ClimateNA_v7.21.exe"
#files <- list.files(pattern = '*.csv$')
#histClimateNA(files[1],'1981','M',exe)

## Decadal climate data
#setwd("E:/Climatena_v721");getwd() # Set up location for the application
#exe <- "E:/Climatena_v721/ClimateNA_v7.21.exe"
#files <- list.files(pattern = '*.csv$')
#histClimateNA(files[1],'1981_1990','M',exe)

## 30 years normals
#setwd("E:/Climatena_v721");getwd() # Set up location for the application
#exe <- "E:/Climatena_v721/ClimateNA_v7.21.exe"
#files <- list.files(pattern = '*.csv$')
#histClimateNA(files[1],'1991_2020','M',exe)
}
\author{
Michael Burnett - UBC
}
