% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprojectASCII.R
\name{reprojectASCII}
\alias{reprojectASCII}
\title{Reproject ASCII raster file}
\usage{
reprojectASCII(file, direc = getwd(), NAflag = -9999, projection = "EPSG:4326")
}
\arguments{
\item{file}{string. ASCII file that requires reprojection}

\item{direc}{The output directory.}

\item{NAflag}{NA value within the raster to mask when reprojecting.}

\item{projection}{string. The output projection in which the ASCII input requires to be reprojected to. The default value is "EPSG:4326"}
}
\value{
Same ASCII file reprojected
}
\description{
Reads and reprojects an ASCII raster file following the data acquisition with ClimateNA (or any alternative source).
}
\examples{
files <- list.files(pattern='*asc$')
for(i in files){
  reprojectASCII(i,NAflag = 32647)
}
}
\author{
Michael Burnett - UBC
}
